Move ResultSet class into Flask app code ...which simplifies config somewhat - Also added words to README (should probably edit for brevity later).
