+++ .gitignore
@@ -2,7 +2,8 @@
 *venv
 
 # Ignore generated files
-*.pyc
+*.py[cod]
+__pycache__/
 
 # Ignore other specific files
 .project
